Added Get-NewDomainUsersGroups Module
This module enumerates domain user or group accounts that were created
in the last certain number of days. The module can also filter the
results for only domain user accounts that are not members of the domain
admins security group or for only domain user accounts that are members
of the domain admins security group.